优化八邻域ing，使得左右边界搜索到尽头的中间时不再搜索候选点，而是等待另一边搜索过来直到相遇在中间，我在写这段话的时候突然意识到我写错了，先commit一下 稍后我会继续commit一下

+++ project/code/image.h
@@ -34,8 +34,8 @@ extern uint8_t  ZhongZhi0;
 extern uint8_t  ZhongZhi1;
 extern uint8_t  ZhongZhi;
 extern uint8_t  Image_Down; //图像下移距离
-extern uint32_t Sum_QuanZhong;
-extern uint32_t Sum_ZhongZhi;
+extern uint16_t Sum_QuanZhong;
+extern uint16_t Sum_ZhongZhi;
 extern char     pid_flag;
